Move consumedir entrypoint to docker and fix compose setup
To make it work as before, some env variables are necessary to be
available to the consumedir container. It's only needed there, but
since the env file exists, they are now listed there.

Also partially reverts the consumedir.dockerfile. Using the base-image
makes sure that all build images are created from the same sources.